# apps/web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OnPoint Agent Commerce API",
    "version": "1.0.0",
    "description": "Discover real curator inventory, request fit signals, and complete Celo x402 commerce. Paid endpoints return HTTP 402 before any payment is authorized. OKX A2MCP facade endpoints (/okx/*) settle in USD₮0 on XLayer."
  },
  "servers": [
    { "url": "https://api.onpoint.famile.xyz" }
  ],
  "paths": {
...
    "/api/agent/try-on": {
      "post": {
        "summary": "Request a paid fit render (Celo)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["curatorSlug", "listingId", "photoData"], "properties": { "curatorSlug": { "type": "string" }, "listingId": { "type": "string" }, "photoData": { "type": "string", "description": "data:image/... base64 payload" }, "paymentTxHash": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Try-on render and fit signal" }, "402": { "description": "cUSD or gasless USDC payment requirements" } }
      }
    },
...
    "/api/looks": {
      "get": {
        "summary": "List live looks with optional filters",
        "parameters": [
          { "name": "curator", "in": "query", "schema": { "type": "string" }, "description": "Filter by curator slug" },
          { "name": "tag", "in": "query", "schema": { "type": "string" }, "description": "Filter by tag (array contains)" },
          { "name": "agent", "in": "query", "schema": { "type": "string" }, "description": "Filter by agent wallet address" },
          { "name": "category", "in": "query", "schema": { "type": "string" }, "description": "Filter by metadata.category (e.g. streetwear, casual, formal, event, sport, vintage, ankara, sustainable)" },
          { "name": "occasion", "in": "query", "schema": { "type": "string" }, "description": "Filter by metadata.occasion (e.g. casual, formal, event, sport, outdoor, travel, date-night)" },
          { "name": "season", "in": "query", "schema": { "type": "string" }, "description": "Filter by metadata.season (e.g. spring, summer, fall, winter, all-season)" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 }, "description": "Max results (capped at 100)" }
        ],
        "responses": { "200": {
          "description": "List of looks with resolved images and metadata",
          "content": { "application/json": { "schema": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/LookListItem" }
          } } }
        } }
      },
      "post": {
        "summary": "Create a new look (agent or curator auth)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["title", "listingIds"], "properties": {
          "title": { "type": "string", "description": "Look title (1-120 chars)" },
          "description": { "type": "string", "description": "Optional description / styling notes" },
          "listingIds": { "type": "array", "items": { "type": "string" }, "description": "Listing UUIDs from a curator storefront (2-10 items)", "minItems": 2, "maxItems": 10 },
          "heroListingId": { "type": "string", "description": "UUID of the hero piece (must be in listingIds). Defaults to first." },
          "tags": { "type": "array", "items": { "type": "string" }, "description": "Free-form tags for discovery" },
          "coverImage": { "type": "string", "description": "data:image/jpeg;base64,... (optional, max 5MB)" },
          "curatorSlug": { "type": "string", "description": "Required for curator auth" },
          "status": { "type": "string", "enum": ["live", "draft"], "default": "live", "description": "Save as draft to publish later" }
        } } } } },
        "responses": {
          "201": {
            "description": "Created look with slug, id, collageUrl, and metadata",
            "content": { "application/json": { "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "format": "uuid" },
                "slug": { "type": "string", "description": "URL-safe slug, unique" },
                "title": { "type": "string" },
                "collageUrl": { "type": "string", "format": "uri", "nullable": true },
                "coverImageUrl": { "type": "string", "format": "uri", "nullable": true },
                "metadata": { "$ref": "#/components/schemas/LookMetadata" },
                "status": { "type": "string", "enum": ["live", "draft"] },
                "items": { "type": "array", "items": { "$ref": "#/components/schemas/LookItem" } }
              }
            } } }
          },
          "401": { "description": "Authentication required — see x-agent-address or x-curator-slug headers" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "LookMetadata": {
        "type": "object",
        "properties": {
          "category": { "type": "string", "enum": ["streetwear", "casual", "formal", "event", "sport", "vintage", "ankara", "sustainable"], "nullable": true },
          "occasion": { "type": "string", "enum": ["casual", "formal", "event", "sport", "outdoor", "travel", "date-night"], "nullable": true },
          "season": { "type": "string", "enum": ["spring", "summer", "fall", "winter", "all-season"], "nullable": true }
        },
        "description": "AI-classified metadata for filtering and discovery"
      },
      "LookItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "curatorSlug": { "type": "string" },
          "imageUrl": { "type": "string", "format": "uri", "nullable": true },
          "isHero": { "type": "boolean" },
          "sizes": { "type": "array", "items": { "type": "object", "properties": {
            "size": { "type": "string" },
            "stock": { "type": "integer" },
            "price": { "type": "number", "description": "Price in KES" }
          } } },
          "kit": { "type": "object", "nullable": true, "properties": {
            "club": { "type": "string" },
            "kitType": { "type": "string" },
            "brand": { "type": "string" },
            "season": { "type": "string" }
          } }
        }
      },
      "LookListItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "slug": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "curatorSlug": { "type": "string", "nullable": true },
          "coverImageUrl": { "type": "string", "format": "uri", "nullable": true },
          "collageUrl": { "type": "string", "format": "uri", "nullable": true },
          "heroImageUrl": { "type": "string", "format": "uri", "nullable": true },
          "tags": { "type": "array", "items": { "type": "string" } },
          "metadata": { "$ref": "#/components/schemas/LookMetadata" },
          "tryOnCount": { "type": "integer" },
          "shareCount": { "type": "integer" },
          "itemCount": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}